feat: Implement signup functionality in Signup use case
This commit adds the implementation of the `call` method in the `Signup` use case class. The `call` method calls the `signup` method of the `AuthRepository` to create a new user with the provided signup parameters. It returns a `UserModel` object representing the newly created user.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.