feat: Update LoginPage to import 'firebase_auth' package
This commit updates the import statement in the LoginPage file to import the 'firebase_auth' package. This change is necessary to use the FirebaseAuth instance for user authentication in the login process.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.