feat: Add signup use case to AuthBloc
This commit adds the signup use case to the AuthBloc class in the auth_bloc.dart file. The signup use case is required for handling user signup functionality. This feature enhances the functionality of the AuthBloc by allowing users to sign up for an account.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.