feat: Add signup functionality to AuthBloc
This commit adds the signup functionality to the AuthBloc class in the auth_bloc.dart file. The signup functionality allows users to sign up for an account by calling the signup use case and handling the response. This feature enhances the functionality of the AuthBloc and improves the user experience.

Note: This commit message follows the established convention of using a prefix (feat) to indicate a new feature or enhancement.